+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.2.0] - 2019-11-25
+
+### Added
+
+- OpaqueSmartID class
+  - Add `idp_tag_blacklist` option
+
+### Changed
+
+- OpaqueSmartID class
+  - Replace `skip_tag_list` with `idp_tag_whitelist` option
+
 ## [v2.1.0] - 2019-09-18
 
 ### Added